# src/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/recipes": {
      "get": {
        "description": "Fetches all recipes in the system. When `lang` is specified, projects localized title and description according to preference hierarchy.",
        "operationId": "RecipeController_getRecipes",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Target locale code for projected titles and descriptions (it, en, fr)",
            "schema": {
              "enum": [
                "it",
                "en",
                "fr"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of recipes with translations"
          }
        },
        "summary": "List all recipes",
        "tags": [
          "Recipes"
        ]
      },
      "post": {
        "description": "Creates a recipe in the declared `source_lang`, verifies language via LibreTranslate, validates catalog references, generates automated multilingual translations for all supported locales, and stores relational records.",
        "operationId": "RecipeController_addRecipe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizedRecipeDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed on inputs or referenced catalog items do not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: insufficient permissions for this action on the target resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create a new recipe",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/page": {
      "get": {
        "description": "Fetches a 30-item page window of recipes optimized for infinite scrolling or stacked presentation. Includes author username and pagination indicators.",
        "operationId": "RecipeController_getRecipeStack",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page index number (1-based, default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "value",
            "required": false,
            "in": "query",
            "description": "Legacy page parameter alias",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated recipe window with hasNextPage and hasPreviousPage booleans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipePaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get paginated recipe stack",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/search": {
      "get": {
        "description": "Performs a case-insensitive search across recipe translation titles and returns matching recipes.",
        "operationId": "RecipeController_getRecipe",
        "parameters": [
          {
            "name": "value",
            "required": true,
            "in": "query",
            "description": "Text query to match in recipe titles",
            "schema": {
              "example": "carbonara",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recipes matching the title search query"
          }
        },
        "summary": "Search recipes by title",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "description": "Fetches full recipe details including localized title, description, storage tips, ordered steps with illustrations, categorized ingredients, tags, and media gallery.",
        "operationId": "RecipeController_getRecipeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique numeric recipe ID",
            "schema": {
              "example": 42,
              "type": "number"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Target localization language (it, en, fr). Defaults to author source_lang if omitted or untranslated.",
            "schema": {
              "enum": [
                "it",
                "en",
                "fr"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete localized recipe specification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalizedRecipeDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get recipe by ID",
        "tags": [
          "Recipes"
        ]
      },
      "patch": {
        "description": "Modifies recipe attributes, steps, ingredients, or localized text. Optionally triggers full re-translation via `retranslate=true`.",
        "operationId": "RecipeController_updateRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Recipe ID",
            "schema": {
              "example": 42,
              "type": "number"
            }
          },
          {
            "name": "retranslate",
            "required": false,
            "in": "query",
            "description": "Set to \"true\" to re-generate automated translations for updated text",
            "schema": {
              "example": "true",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recipe updated successfully"
          },
          "400": {
            "description": "Invalid update payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: insufficient permissions for this action on the target resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update recipe details",
        "tags": [
          "Recipes"
        ]
      },
      "delete": {
        "description": "Permanently deletes the recipe and all its steps, media, translations, and associated likes.",
        "operationId": "RecipeController_deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Recipe ID",
            "schema": {
              "example": 42,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: insufficient permissions for this action on the target resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete recipe",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/{id}/translate": {
      "post": {
        "description": "Re-triggers LibreTranslate for a recipe whose translation status is marked as failed.",
        "operationId": "RecipeController_retryTranslation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Recipe ID",
            "schema": {
              "example": 42,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation retry processed"
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: insufficient permissions for this action on the target resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Retry automated translations",
        "tags": [
          "Recipes"
        ]
      }
    },
...
  },
  "info": {
    "title": "WeCook API",
    "description": "RESTful backend API specification for the WeCook recipe sharing social platform.\n\n### Authentication\nAuthentication is based on secure, HttpOnly JWT cookies:\n- **`accessToken`**: Short-lived (10 minutes) JWT sent automatically by browsers or HTTP clients on protected endpoints.\n- **`refresh_token`**: Long-lived (7 days) JWT restricted to the `/api/auth/refresh` path.\n\n### Media Storage & Static Assets\nUploaded recipe covers, step images, gallery photos, videos, and user avatars are served statically under `/uploads/`.\n\n### Localization & Multilingual Engine\nRecipes and catalogs support Italian (`it`), English (`en`), and French (`fr`), powered by LibreTranslate relational projection.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://app.we-cook.it:8443",
      "description": "Staging / Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT Refresh Token stored in HttpOnly cookie (`refresh_token`) for the refresh endpoint"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code representing the error category",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 formatted timestamp of when the error occurred",
            "example": "2026-10-09T17:30:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Requested URL path that produced the error",
            "example": "/api/recipes/42"
          },
          "message": {
            "description": "Error description or list of validation error messages",
            "example": "Recipe not found",
            "oneOf": [
              {
                "type": "string",
                "example": "Recipe not found"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "Title must be at least 3 characters long"
                ]
              }
            ]
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message"
        ]
      },
...
    }
  }
}